# ChurrasAPI/apiweb.churras.show/apiweb.churras.show/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API ChurrasShow",
    "description": "API Para Gerenciamento de Eventos e Or�amento - ChurrasShow WEB.API",
    "contact": {
      "name": "Alunos - Senai Inform�tica",
      "url": "https://github.com/BrizidoAndre/ChurrasShow"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Usuario": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Nome",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Email",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Senha",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Rg",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Cpf",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Logradouro",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Cidade",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Uf",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Cep",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Numero",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Bairro",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Complemento",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "IdTipoUsuario",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "formData",
            "name": "Foto",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Arquivo",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Usuario"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/AlterarSenha": {
      "put": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AlterarSenhaViewModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/AtualizarUsuario": {
      "put": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AtualizarUsuarioViewModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/AlterarFotoPerfil": {
      "put": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "formData",
            "name": "Arquivo",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "Foto",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/{id}": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "AlterarSenhaViewModel": {
      "required": [
        "senhaNova"
      ],
      "type": "object",
      "properties": {
        "senhaNova": {
          "minLength": 1,
          "type": "string"
        }
      }
    },
...
    "AtualizarUsuarioViewModel": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "rg": {
          "type": "string"
        },
        "cpf": {
          "type": "string"
        },
        "logradouro": {
          "type": "string"
        },
        "cidade": {
          "type": "string"
        },
        "uf": {
          "type": "string"
        },
        "cep": {
          "format": "int32",
          "type": "integer"
        },
        "number": {
          "format": "int32",
          "type": "integer"
        },
        "bairro": {
          "type": "string"
        },
        "complemento": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insira o token JWT desta maneira: Bearer {seu token}"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}